Flink: move the classes from flink.sink.shuffle.statistics pkg to one level up as flink.sink.shuffle pkg
