+++ flink/v1.17/flink/src/main/java/org/apache/iceberg/flink/sink/shuffle/MapDataStatisticsFactory.java
@@ -16,18 +16,19 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iceberg.flink.sink.shuffle.statistics;
+package org.apache.iceberg.flink.sink.shuffle;
 
 import org.apache.flink.annotation.Internal;
 
 /**
- * DataStatisticsFactory defines the interface to create {@link DataStatistics}.
- *
- * <p>For low-cardinality key, MapDataStatisticsFactory will be implemented to create
- * MapDataStatistics.
+ * MapDataStatisticsFactory creates {@link MapDataStatistics} to track traffic volume for
+ * low-cardinality key in hash mode
  */
 @Internal
-public interface DataStatisticsFactory<K> {
+class MapDataStatisticsFactory<K> implements DataStatisticsFactory<K> {
 
-  DataStatistics<K> createDataStatistics();
+  @Override
+  public DataStatistics<K> createDataStatistics() {
+    return new MapDataStatistics<>();
+  }
 }
