+++ flink/v1.17/flink/src/test/java/org/apache/iceberg/flink/sink/shuffle/TestDataStatisticsOperator.java
@@ -47,10 +47,6 @@ import org.apache.flink.streaming.runtime.tasks.StreamMockEnvironment;
 import org.apache.flink.streaming.util.MockOutput;
 import org.apache.flink.streaming.util.MockStreamConfig;
 import org.apache.flink.streaming.util.OneInputStreamOperatorTestHarness;
-import org.apache.iceberg.flink.sink.shuffle.statistics.DataStatistics;
-import org.apache.iceberg.flink.sink.shuffle.statistics.DataStatisticsFactory;
-import org.apache.iceberg.flink.sink.shuffle.statistics.MapDataStatistics;
-import org.apache.iceberg.flink.sink.shuffle.statistics.MapDataStatisticsFactory;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableList;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 import org.junit.After;
